Split cli args into sub-modules
