+++ src/nhp_dwiproc/app/__init__.py
@@ -1,7 +1,7 @@
 """Initialize application module."""
 
 from nhp_dwiproc.app import analysis_levels, type
-from nhp_dwiproc.app.cli import parser
+from nhp_dwiproc.app.cli.parser import parser
 from nhp_dwiproc.app.descriptor import generate_descriptor
 from nhp_dwiproc.app.utils.app import initialize, validate_cfg
